requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-axes==8.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.4
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
gunicorn==21.2.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-axes==8.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.4
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
gunicorn==21.2.0
django-axes==8.0.0
django-filter==24.1
dj-database-url
